Stock price calculation nf

+++ StockValue-Service/src/main/java/ie/atu/stockvalueservice/StockValueController.java
@@ -16,7 +16,7 @@ public class StockValueController {
     }
 
     @GetMapping("/getStockVal/{stockName}")
-    public int portfolioFromStockVal(@PathVariable String stockName){
+    public double portfolioFromStockVal(@PathVariable String stockName){
         return stockValueService.portfolioResult(stockName);
     }
 }
